force to double precision before interpolating
git-svn-id: https://sccn.ucsd.edu/svn/software/eeglab@9555 1e58238c-3a72-4c0f-aabe-36d030380f39

+++ functions/sigprocfunc/topoplot.m
@@ -954,8 +954,8 @@ if ~strcmpi(STYLE,'blank') % if draw interpolated scalp map
   yi = linspace(ymin,ymax,GRID_SCALE);   % y-axis description (row vector)
 
   try
-      [Xi,Yi,Zi] = griddata(inty,intx,intValues,yi',xi,'v4'); % interpolate data
-      [Xi,Yi,ZiC] = griddata(inty,intx,intContourVals,yi',xi,'v4'); % interpolate data
+      [Xi,Yi,Zi] = griddata(inty,intx,double(intValues),yi',xi,'v4'); % interpolate data
+      [Xi,Yi,ZiC] = griddata(inty,intx,double(intContourVals),yi',xi,'v4'); % interpolate data
   catch,
       [Xi,Yi,Zi] = griddata(inty,intx,intValues',yi,xi'); % interpolate data (Octave)
       [Xi,Yi,ZiC] = griddata(inty,intx,intContourVals',yi,xi'); % interpolate data
